Moved creation of agg and vs to support_files.py
