+++ pxbuild/controll/from_pxmetadata_file.py
@@ -7,20 +7,18 @@ from typing import List, Dict
 from pxbuild.models.input.pydantic_pxmetadata import PxMetadata
 from pxbuild.models.input.pydantic_pxbuildconfig import PxbuildConfig
 from pxbuild.models.input.pydantic_pxcodes import PxCodes
-from pxbuild.models.input.pydantic_pxcodes import Grouping
 from pxbuild.models.input.helper_pxcodes import HelperPxCodes
 from pxbuild.models.input.pydantic_pxstatistics import PxStatistics
 
 from pxbuild.models.output.pxfile.px_file_model import PXFileModel
-from pxbuild.models.output.agg_vs.vs_file_model import _VSFileModel
-from pxbuild.models.output.agg.agg_file_model import AggFileModel
 
 from .helpers.small_static_functions import Commons
 
 from .helpers.datadata_helpers.datadatasource import Datadatasource
 from .helpers.datadata_helpers.for_get_data import ForGetData
 from .helpers.datadata_helpers.main_data import MapData
-from .helpers.loaded_jsons import LoadedJsons 
+from .helpers.loaded_jsons import LoadedJsons
+from .helpers.support_files import SupportFiles 
 
 
 class LoadFromPxmetadata:
@@ -104,7 +102,8 @@ class LoadFromPxmetadata:
             write_output(self._pxmetadata_id, self._config.admin.output_destination.px_folder_format, out_model)
             self.models_for_pytest["multi"] = out_model 
 
-        self.make_vs_file()
+        support = SupportFiles(self._pxmetadata_model,self._config,self._resolved_pxcodes_ids,self._pxcodes_helper,self._pxmetadata_id)
+        support.make_vs_file()
 
 
 
@@ -190,7 +189,7 @@ class LoadFromPxmetadata:
                 temp_values = my_pxcodes_helper.getLabels(lang)
                 out_model.values.set(temp_values, my_funny_var_id, lang)
                 if my_codes.groupings:
-                    my_domain_id = make_domain_id(my_var.codelist_id, self._current_lang)
+                    my_domain_id = Commons.make_domain_id(my_var.codelist_id, self._current_lang)
                     out_model.domain.set(my_domain_id, my_funny_var_id, self._current_lang)
 
                 if my_var.is_geo_variable_type:
@@ -365,90 +364,6 @@ class LoadFromPxmetadata:
         out_model.source.set(in_config.source[self._current_lang], self._current_lang)
 
 
-    #################################
-    # def makeVsFile(self,out_vs_model:_VSFileModel):
-    def make_vs_file(self):
-        if not self._pxmetadata_model.dataset.coded_dimensions:
-            return
-        
-        for language in self._config.admin.valid_languages:
-
-            for my_var in self._pxmetadata_model.dataset.coded_dimensions:
-
-                out_vs_model = _VSFileModel()
-                my_codes: PxCodes = self._resolved_pxcodes_ids[my_var.codelist_id]
-                if my_codes.groupings:
-                    vs_name = make_domain_id(my_var.codelist_id, language)
-                    #vs_type = "G" if my_var.is_geo_variable_type else "V"
-                    vs_type = "V" # TODO type could be V,H or N
-                    out_vs_model.description.set("Name", vs_name)
-                    out_vs_model.description.set("Type", vs_type)
-
-                    my_pxcodes_helper = self._pxcodes_helper[my_var.codelist_id]
-                    group_conter = 0
-                    for groups in my_codes.groupings:
-                        group_conter = group_conter + 1
-                        group_key = str(group_conter)
-                        out_vs_model.aggreg.set(group_key, groups.filename_base + "_" + language + ".agg")
-                        self.make_agg_file(groups, vs_name, my_pxcodes_helper, language)
-                    my_domain = make_domain_id(my_var.codelist_id, language)
-                    out_vs_model.domain.set("1", my_domain)
-
-                    
-                    value_item_counter = 0
-                    for my_item in my_pxcodes_helper._sorted_valueitems[language]:
-                        value_item_counter = value_item_counter + 1
-                        value_item_key = str(value_item_counter)
-                        my_stripped_code = my_item.code.strip("'")
-                        out_vs_model.valuecode.set(value_item_key, my_stripped_code)
-                        my_stripped_text = my_item.label[language].strip("'")
-                        out_vs_model.valuetext.set(value_item_key, my_stripped_text)
-
-                    out_folder_format: str = self._config.admin.output_destination.agg_folder_format
-                    out_folder = out_folder_format.format(id=self._pxmetadata_id)
-                    out_file = out_folder + "/" + my_codes.id + "_" + language + ".vs"
-
-                    with open(out_file, "w") as f:
-                        print(out_vs_model, file=f)
-                        print("File written to:", out_file)
-
-    def make_agg_file(self, grouping: Grouping, vs_name: str, my_pxcodes_helper:HelperPxCodes, language:str):
-        out_agg_model = AggFileModel()
-        #aggreg_name = grouping.filename_base + "_" + language
-        aggreg_name = grouping.label[language]
-        out_agg_model.set("Aggreg", "Name", aggreg_name)
-        out_agg_model.set("Aggreg", "Valueset", vs_name)
-        item_counter = 0
-        # section= "Aggreg"
-        for item in grouping.valueitems:
-            item_counter = item_counter + 1
-            item_key = str(item_counter)
-            groupcode = item.code
-            valuetext = item.label[language]
-            out_agg_model.set("Aggreg", item_key, groupcode)
-            out_agg_model.set("Aggtext", item_key, valuetext)
-
-            child_code_conter = 0
-            #ordered_children = [code for code in my_pxcodes_helper.getCodes(language) if item.unordered_children and code in item.unordered_children]
-
-            ordered_children:List[str] = []
-            for code in my_pxcodes_helper.getCodes(language):
-                if item.unordered_children and code in item.unordered_children:
-                   ordered_children.append(code) 
-            
-            for child_code in ordered_children:
-                child_code_conter = child_code_conter + 1
-                child_code_key = str(child_code_conter)
-                out_agg_model.set(groupcode, child_code_key, child_code)
-
-        out_folder_format: str = self._config.admin.output_destination.agg_folder_format
-        out_folder = out_folder_format.format(id=self._pxmetadata_id)
-        out_file = out_folder + "/"  + str(grouping.filename_base) + "_" + language + ".agg"
-        with open(out_file, "w") as f:
-            print(out_agg_model, file=f)
-            print("File written to:", out_file)
-        print("i agg")
-        print(grouping.filename_base)
 
 
 def convert_to_pxdate_string(date_string: str, date_format: str) -> str:
@@ -456,11 +371,6 @@ def convert_to_pxdate_string(date_string: str, date_format: str) -> str:
     px_date_string = dtm_date.strftime(f"%Y%m%d %H:%M")
 
     return px_date_string
-
-
-def make_domain_id(code_list_id: str, lang: str) -> str:
-    domain_id = code_list_id + "_" + lang
-    return domain_id
 
 def write_output(pxmetadata_id:str, px_folder_format:str, out_model:PXFileModel, language:str | None = None) -> None:
     temp_tabid = pxmetadata_id
