+++ pxbuild/controll/helpers/small_static_functions.py
@@ -5,3 +5,9 @@ class Commons:
   @staticmethod
   def get_variable_list(stub:List[str], heading:List[str] ) -> List[str]:
         return stub + heading
+  
+
+  @staticmethod
+  def make_domain_id(code_list_id: str, language: str) -> str:
+    domain_id = code_list_id + "_" + language
+    return domain_id
